Create client module to avoid circular import.
