+++ topshelfsoftware_aws_util/sfn.py
@@ -5,7 +5,7 @@ import json
 import time
 import uuid
 
-from topshelfsoftware_aws_util import create_boto3_client
+from topshelfsoftware_aws_util.client import create_boto3_client
 from topshelfsoftware_util.json import fmt_json
 from topshelfsoftware_util.log import get_logger
