+++ topshelfsoftware_aws_util/ssm.py
@@ -2,7 +2,7 @@
 
 from botocore.exceptions import ClientError as BotoClientError
 
-from topshelfsoftware_aws_util import create_boto3_client
+from topshelfsoftware_aws_util.client import create_boto3_client
 from topshelfsoftware_util.json import fmt_json
 from topshelfsoftware_util.log import get_logger
